Add a PeriodicPingHandler and a PingListenHandler
