+++ protocols/ping/Cargo.toml
@@ -5,6 +5,7 @@ authors = ["Parity Technologies <admin@parity.io>"]
 license = "MIT"
 
 [dependencies]
+arrayvec = "0.4"
 bytes = "0.4"
 libp2p-core = { path = "../../core" }
 log = "0.4.1"
@@ -15,6 +16,8 @@ parking_lot = "0.6"
 rand = "0.5"
 tokio-codec = "0.1"
 tokio-io = "0.1"
+tokio-timer = "0.2.6"
+void = "1.0"
 
 [dev-dependencies]
 libp2p-tcp-transport = { path = "../../transports/tcp" }
